refactor: organised model files, changed re-exports
