+++ src/common/camera/mod.rs
@@ -1,6 +1,7 @@
-pub mod camera;
+mod camera;
 mod fps_camera;
 mod orbital_camera;
 
+pub use camera::Camera;
 pub use fps_camera::FpsCamera;
 pub use orbital_camera::OrbitalCamera;
