+++ src/common/lib.rs
@@ -6,8 +6,7 @@ pub mod debug;
 pub mod input;
 pub mod line;
 pub mod maths;
-pub mod model;
-pub mod model_instance;
+pub mod models;
 pub mod renderer;
 pub mod scene;
 pub mod serde;
