+++ src/common/models/mod.rs
@@ -1,0 +1,5 @@
+mod model;
+mod model_instance;
+
+pub use model::Model;
+pub use model_instance::ModelInstance;
